Address .append deprecation warning
Using pandas==1.4.2 (matching glacio01 version) I get deprecation warnings for .append method,
stating that .append should be replaced by .concat. However, .concat is not a method on a dataframe
so we cannot replace directly. Instead I first build up the mirrored dataframe, then proceed with
the windowing, etc as before. This also should be more readable.

Note that I compared the returned array output with both the previous method and this method,
using np.array_equal(), and the two methods are identical.