implement writing of position csv file to aws-l3
This uses already built-in building of positions dict to write a csv file to the aws-l3 repo.
The positions are the most recent point on 3 month linear regression trends.
There is some inconsistency, in that we only look for and use modem positions when GPS positions are missing
for stations that make it through to BUFR submission. There are many scenarios in getBUFR where a station is
skipped or rejected for BUFR processing. I have captured all of these situations and we will still write position
metadata to the csv file for these stations (but only using GPS, not modem).

This will require to run as 'getBUFR --positions' in processing script.