persist uid in ini file (#77)
Modification to getL0tx will write the previous last_uid back to the ini file if no new messages are found. The modification to tx.py is simply to improve the print statement. The last_uid should not be included in the list of new UIDs that is printed. As commented in the code, the search command always returns at least the most recent message, even if it has already been processed. So just remove from the print (it is also skipped in the processing).

Resolves: #76