fixed resampling for precipitation
The hourly, daily, monthly resampling was not working for precipitation.

Since precip_*_cor should be at this stage an only accumulating quantity at high temporal resolution (10 min and hourly), we now need, at daily and monthly time scale, to sample the hi-res data at the daily or monthly timestamps.

The precipitation rate is then calculated as the increments that are seen between these sample timestamps. For the first time interval, which is likely incomplete, we look at how many values were used and if there are sufficient, we still define a precipitation rate from the (incomplete) accumulated precipitation.

The same approach could also be implemented for the last time interval, but it is less critical since new data will eventually come to complete that latest daily or monthly interval.

In addtion to these modification, some variables were renamed to be more in line with what they are.